+++ allennlp/commands/__init__.py
@@ -5,6 +5,7 @@ from typing import Any, Optional
 from overrides import overrides
 
 from allennlp import __version__
+from allennlp.commands.build_vocab import BuildVocab
 from allennlp.commands.evaluate import Evaluate
 from allennlp.commands.find_learning_rate import FindLearningRate
 from allennlp.commands.predict import Predict