+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Added a `build-vocab` subcommand that can be used to build a vocabulary from a training config file.
 - Added a script: `scripts/release_notes.py`, which automatically prepares markdown release notes from the
   CHANGELOG and commit history.
 - Added a flag `--predictions-output-file` to the `evaluate` command, which tells AllenNLP to write the